+++ p2p/module.go
@@ -43,11 +43,17 @@ type p2pModule struct {
 	identity       libp2p.Option
 	listenAddrs    libp2p.Option
 
+	// TECHDEBT(#810): register the providers to the module registry instead of
+	// holding a reference in the module struct and passing via router config.
+	//
 	// Assigned during creation via `#setupDependencies()`.
 	currentHeightProvider providers.CurrentHeightProvider
 	pstoreProvider        providers.PeerstoreProvider
 	nonceDeduper          *mempool.GenericFIFOSet[uint64, uint64]
 
+	// TECHDEBT(#810): register the routers to the module registry instead of
+	// holding a reference in the module struct. This will improve testability.
+	//
 	// Assigned during `#Start()`. TLDR; `host` listens on instantiation.
 	// and `router` depends on `host`.
 	router typesP2P.Router
@@ -252,6 +258,9 @@ func (m *p2pModule) setupPeerstoreProvider() error {
 	if !ok {
 		return fmt.Errorf("unknown peerstore provider type: %T", pstoreProviderModule)
 	}
+
+	// TECHDEBT(#810): register the provider to the module registry instead of
+	// holding a reference in the module struct and passing via router config.
 	m.pstoreProvider = pstoreProvider
 
 	return nil
@@ -260,6 +269,7 @@ func (m *p2pModule) setupPeerstoreProvider() error {
 // setupCurrentHeightProvider attempts to retrieve the current height provider
 // from the bus registry, falls back to the consensus module if none is registered.
 func (m *p2pModule) setupCurrentHeightProvider() error {
+	// TECHDEBT(#810): simplify once submodules are more convenient to retrieve.
 	m.logger.Debug().Msg("setupCurrentHeightProvider")
 	currentHeightProviderModule, err := m.GetBus().GetModulesRegistry().GetModule(current_height_provider.ModuleName)
 	if err != nil {
@@ -276,6 +286,9 @@ func (m *p2pModule) setupCurrentHeightProvider() error {
 	if !ok {
 		return fmt.Errorf("unexpected current height provider type: %T", currentHeightProviderModule)
 	}
+
+	// TECHDEBT(#810): register the provider to the module registry instead of
+	// holding a reference in the module struct and passing via router config.
 	m.currentHeightProvider = currentHeightProvider
 
 	return nil
@@ -294,6 +307,8 @@ func (m *p2pModule) setupNonceDeduper() error {
 
 // setupRouter instantiates the configured router implementation.
 func (m *p2pModule) setupRouter() (err error) {
+	// TECHDEBT(#810): register the router to the module registry instead of
+	// holding a reference in the module struct. This will improve testability.
 	m.router, err = raintree.NewRainTreeRouter(
 		m.GetBus(),
 		&config.RainTreeConfig{